fix location lat and lng

+++ backend/migrations/20211116132933-create-location.js
@@ -9,10 +9,10 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       lat: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(8,6)
       },
       lng: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(9,6)
       },
       address: {
         type: Sequelize.STRING
